Remove v28 references from sbt Let all others use Play version 3.
detail:https://github.com/guardian/play-secret-rotation/issues/461
