+++ build.sbt
@@ -3,6 +3,7 @@ import sbtversionpolicy.withsbtrelease.ReleaseVersion.fromAggregatedAssessedComp
 
 lazy val baseSettings = Seq(
   scalaVersion := "2.13.14",
+  crossScalaVersions := Seq(scalaVersion.value, "3.3.3"),
   organization := "com.gu.play-secret-rotation",
   licenses := Seq("Apache-2.0" -> url("https://www.apache.org/licenses/LICENSE-2.0")),
   scalacOptions ++= Seq("-deprecation", "-unchecked", "-release:11"),
@@ -10,10 +11,8 @@ lazy val baseSettings = Seq(
     Tests.Argument(TestFrameworks.ScalaTest,"-u", s"test-results/scala-${scalaVersion.value}")
 )
 
-lazy val crossCompileScala3 = crossScalaVersions := Seq(scalaVersion.value, "3.3.3")
-
 lazy val core =
-  project.settings(crossCompileScala3, baseSettings).settings(
+  project.settings(baseSettings).settings(
     libraryDependencies ++= Seq(
       "com.github.blemale" %% "scaffeine" % "5.2.1",
       "com.typesafe.scala-logging" %% "scala-logging" % "3.9.5",
@@ -23,7 +22,7 @@ lazy val core =
   )
 
 lazy val `aws-parameterstore-secret-supplier-base` =
-  project.in(file("aws-parameterstore/secret-supplier")).settings(crossCompileScala3, baseSettings).dependsOn(core)
+  project.in(file("aws-parameterstore/secret-supplier")).settings(baseSettings).dependsOn(core)
 
 val awsSdkForVersion = Map(
   1 -> "com.amazonaws" % "aws-java-sdk-ssm" % "1.12.716",
@@ -32,7 +31,7 @@ val awsSdkForVersion = Map(
 
 def awsParameterStoreWithSdkVersion(version: Int)=
   Project(s"aws-parameterstore-sdk-v$version", file(s"aws-parameterstore/secret-supplier/aws-sdk-v$version"))
-  .settings(crossCompileScala3, baseSettings)
+  .settings(baseSettings)
   .dependsOn(`aws-parameterstore-secret-supplier-base`)
   .settings(libraryDependencies += awsSdkForVersion(version))
 
@@ -40,7 +39,7 @@ lazy val `aws-parameterstore-sdk-v1` = awsParameterStoreWithSdkVersion(1)
 lazy val `aws-parameterstore-sdk-v2` = awsParameterStoreWithSdkVersion(2)
 
 lazy val `aws-parameterstore-lambda` = project.in(file("aws-parameterstore/lambda"))
-  .settings(crossCompileScala3, baseSettings).dependsOn(`secret-generator`).settings(
+  .settings(baseSettings).dependsOn(`secret-generator`).settings(
   libraryDependencies ++= Seq(
     "com.amazonaws" % "aws-lambda-java-core" % "1.2.3",
     "com.amazonaws" % "aws-lambda-java-events" % "3.11.5",
@@ -48,10 +47,9 @@ lazy val `aws-parameterstore-lambda` = project.in(file("aws-parameterstore/lambd
   )
 )
 
-lazy val `secret-generator` = project.settings(crossCompileScala3, baseSettings)
+lazy val `secret-generator` = project.settings(baseSettings)
 
 val exactPlayVersions = Map(
-  "28" -> "com.typesafe.play" %% "play" % "2.8.21",
   "29" -> "com.typesafe.play" %% "play" % "2.9.2",
   "30" -> "org.playframework" %% "play" % "3.0.2"
 )
@@ -63,16 +61,13 @@ def playVersion(majorMinorVersion: String)= {
     .settings(libraryDependencies += exactPlayVersions(majorMinorVersion))
 }
 
-
-lazy val `play-v28` = playVersion("28")
-lazy val `play-v29` = playVersion("29").settings(crossCompileScala3)
-lazy val `play-v30` = playVersion("30").settings(crossCompileScala3)
+lazy val `play-v29` = playVersion("29")
+lazy val `play-v30` = playVersion("30")
 
 
 lazy val `play-secret-rotation-root` = (project in file("."))
   .aggregate(
     core,
-    `play-v28`,
     `play-v29`,
     `play-v30`,
     `aws-parameterstore-secret-supplier-base`,
